Aztec Diamond V2
This is the latest version, I uploaded the code to work as a matlab application. This allowed for an increase in the output. I currently also added a loading bar so that when I ran the A(1000) I could see the progress. I takes about 5 min for me to run this code. However, the vast majority of the time is at the last end of the computations between the 800 and the 1000th square. Also, the picture is nicely generated into a matlab window and not using excel.